Made Data_Struct.py follow rubric, Probably? I really, really should check to see if that's actually working correctly.
While at it, made init.py to actually use the functions from Data_Struct.py so I could initialize the rudementary blockchain, and updated Main.py to take advantage of it.

Last thing, added Makefile so that it becomes a proper executable, so that submission won't be borked. This is (probably) what is needed for testing. I think. I hope. Aaa I should be asleep.

Be sure to ping me if needed updated enviroment variables -I don't think it's needed but I'm not dismissing the chance it is. Hopefully.